Only add one process.exit handler

+++ process.js
@@ -6,7 +6,7 @@ const bin = './ngrok' + (platform === 'win32' ? '.exe' : '');
 const ready = /starting web service.*addr=(\d+\.\d+\.\d+\.\d+:\d+)/;
 const inUse = /address already in use/;
 
-let processPromise, activeProcess;
+let processPromise, activeProcess, processHasExit;
 
 /*
 	ngrok process runs internal ngrok api
@@ -71,7 +71,10 @@ async function startProcess (opts) {
 		activeProcess = null;
 	});
 
-	process.on('exit', async () => await killProcess());
+        if (!processHasExit) {
+	        process.on('exit', async () => await killProcess());
+                processHasExit = true;
+        }
 
 	try {
 		const url = await apiUrl;
